Create .gitignore
The .gitignore file is a configuration file used by Git to determine which files and directories should be ignored and not tracked by version control. When you're working on a project with Git, there are often files or directories that you don't want to include in the repository. These could be temporary files, build artifacts, user-specific settings, or any other files that are generated during the development process and are not essential to the project's source code.